ffi module cleanup: context.h to frameobject.h
