+++ src/ffi/cpython/mod.rs
@@ -1,11 +1,17 @@
 pub mod abstract_;
+// skipped bytearrayobject.h
 #[cfg(not(PyPy))]
 pub mod bytesobject;
 pub mod ceval;
 pub mod code;
+pub mod dictobject;
+// skipped fileobject.h
+pub mod frameobject;
 
 pub use self::abstract_::*;
 #[cfg(not(PyPy))]
 pub use self::bytesobject::*;
 pub use self::ceval::*;
 pub use self::code::*;
+pub use self::dictobject::*;
+pub use self::frameobject::*;
