+++ src/ffi/dictobject.rs
@@ -6,33 +6,6 @@ use std::os::raw::{c_char, c_int};
 extern "C" {
     #[cfg_attr(PyPy, link_name = "PyPyDict_Type")]
     pub static mut PyDict_Type: PyTypeObject;
-    pub static mut PyDictIterKey_Type: PyTypeObject;
-    pub static mut PyDictIterValue_Type: PyTypeObject;
-    pub static mut PyDictIterItem_Type: PyTypeObject;
-    pub static mut PyDictKeys_Type: PyTypeObject;
-    pub static mut PyDictItems_Type: PyTypeObject;
-    pub static mut PyDictValues_Type: PyTypeObject;
-    #[cfg(Py_3_8)]
-    pub static mut PyDictRevIterKey_Type: PyTypeObject;
-    #[cfg(Py_3_8)]
-    pub static mut PyDictRevIterValue_Type: PyTypeObject;
-    #[cfg(Py_3_8)]
-    pub static mut PyDictRevIterItem_Type: PyTypeObject;
-}
-
-#[repr(C)]
-pub struct PyDictKeysObject {
-    _unused: [u8; 0],
-}
-
-#[repr(C)]
-#[derive(Debug)]
-pub struct PyDictObject {
-    pub ob_base: PyObject,
-    pub ma_used: Py_ssize_t,
-    pub ma_version_tag: u64,
-    pub ma_keys: *mut PyDictKeysObject,
-    pub ma_values: *mut *mut PyObject,
 }
 
 #[inline]
@@ -45,43 +18,14 @@ pub unsafe fn PyDict_CheckExact(op: *mut PyObject) -> c_int {
     (Py_TYPE(op) == &mut PyDict_Type) as c_int
 }
 
-#[inline]
-pub unsafe fn PyDictKeys_Check(op: *mut PyObject) -> c_int {
-    (Py_TYPE(op) == &mut PyDictKeys_Type) as c_int
-}
-
-#[inline]
-pub unsafe fn PyDictItems_Check(op: *mut PyObject) -> c_int {
-    (Py_TYPE(op) == &mut PyDictItems_Type) as c_int
-}
-
-#[inline]
-pub unsafe fn PyDictValues_Check(op: *mut PyObject) -> c_int {
-    (Py_TYPE(op) == &mut PyDictValues_Type) as c_int
-}
-
-#[inline]
-pub unsafe fn PyDictViewSet_Check(op: *mut PyObject) -> c_int {
-    (PyDictKeys_Check(op) != 0 || PyDictItems_Check(op) != 0) as c_int
-}
-
 extern "C" {
     #[cfg_attr(PyPy, link_name = "PyPyDict_New")]
     pub fn PyDict_New() -> *mut PyObject;
-    #[cfg(not(PyPy))]
-    pub fn _PyDict_NewPresized(minused: Py_ssize_t) -> *mut PyObject;
     #[cfg_attr(PyPy, link_name = "PyPyDict_GetItem")]
     pub fn PyDict_GetItem(mp: *mut PyObject, key: *mut PyObject) -> *mut PyObject;
     pub fn PyDict_GetItemWithError(mp: *mut PyObject, key: *mut PyObject) -> *mut PyObject;
     #[cfg_attr(PyPy, link_name = "PyPyDict_SetItem")]
     pub fn PyDict_SetItem(mp: *mut PyObject, key: *mut PyObject, item: *mut PyObject) -> c_int;
-    #[cfg(not(PyPy))]
-    pub fn _PyDict_SetItem_KnownHash(
-        mp: *mut PyObject,
-        key: *mut PyObject,
-        item: *mut PyObject,
-        hash: crate::ffi::Py_hash_t,
-    ) -> c_int;
     #[cfg_attr(PyPy, link_name = "PyPyDict_DelItem")]
     pub fn PyDict_DelItem(mp: *mut PyObject, key: *mut PyObject) -> c_int;
     #[cfg_attr(PyPy, link_name = "PyPyDict_Clear")]
@@ -92,14 +36,6 @@ extern "C" {
         pos: *mut Py_ssize_t,
         key: *mut *mut PyObject,
         value: *mut *mut PyObject,
-    ) -> c_int;
-    #[cfg(not(PyPy))]
-    pub fn _PyDict_Next(
-        mp: *mut PyObject,
-        pos: *mut Py_ssize_t,
-        key: *mut *mut PyObject,
-        value: *mut *mut PyObject,
-        hash: *mut crate::ffi::Py_hash_t,
     ) -> c_int;
     #[cfg_attr(PyPy, link_name = "PyPyDict_Keys")]
     pub fn PyDict_Keys(mp: *mut PyObject) -> *mut PyObject;
@@ -113,8 +49,6 @@ extern "C" {
     pub fn PyDict_Copy(mp: *mut PyObject) -> *mut PyObject;
     #[cfg_attr(PyPy, link_name = "PyPyDict_Contains")]
     pub fn PyDict_Contains(mp: *mut PyObject, key: *mut PyObject) -> c_int;
-    #[cfg(not(PyPy))]
-    pub fn _PyDict_Contains(mp: *mut PyObject, key: *mut PyObject, hash: Py_ssize_t) -> c_int;
     #[cfg_attr(PyPy, link_name = "PyPyDict_Update")]
     pub fn PyDict_Update(mp: *mut PyObject, other: *mut PyObject) -> c_int;
     #[cfg_attr(PyPy, link_name = "PyPyDict_Merge")]
@@ -130,6 +64,60 @@ extern "C" {
     ) -> c_int;
     #[cfg_attr(PyPy, link_name = "PyPyDict_DelItemString")]
     pub fn PyDict_DelItemString(dp: *mut PyObject, key: *const c_char) -> c_int;
-    #[cfg(not(PyPy))]
-    pub fn _PyObject_GetDictPtr(obj: *mut PyObject) -> *mut *mut PyObject;
+// skipped 3.10 / ex-non-limited PyObject_GenericGetDict
 }
+
+#[cfg_attr(windows, link(name = "pythonXY"))]
+extern "C" {
+    pub static mut PyDictKeys_Type: PyTypeObject;
+    pub static mut PyDictValues_Type: PyTypeObject;
+    pub static mut PyDictItems_Type: PyTypeObject;
+}
+
+#[inline]
+pub unsafe fn PyDictKeys_Check(op: *mut PyObject) -> c_int {
+    (Py_TYPE(op) == &mut PyDictKeys_Type) as c_int
+}
+
+#[inline]
+pub unsafe fn PyDictValues_Check(op: *mut PyObject) -> c_int {
+    (Py_TYPE(op) == &mut PyDictValues_Type) as c_int
+}
+
+#[inline]
+pub unsafe fn PyDictItems_Check(op: *mut PyObject) -> c_int {
+    (Py_TYPE(op) == &mut PyDictItems_Type) as c_int
+}
+
+#[inline]
+pub unsafe fn PyDictViewSet_Check(op: *mut PyObject) -> c_int {
+    (PyDictKeys_Check(op) != 0 || PyDictItems_Check(op) != 0) as c_int
+}
+
+#[cfg_attr(windows, link(name = "pythonXY"))]
+extern "C" {
+    pub static mut PyDictIterKey_Type: PyTypeObject;
+    pub static mut PyDictIterValue_Type: PyTypeObject;
+    pub static mut PyDictIterItem_Type: PyTypeObject;
+    #[cfg(Py_3_8)]
+    pub static mut PyDictRevIterKey_Type: PyTypeObject;
+    #[cfg(Py_3_8)]
+    pub static mut PyDictRevIterValue_Type: PyTypeObject;
+    #[cfg(Py_3_8)]
+    pub static mut PyDictRevIterItem_Type: PyTypeObject;
+}
+
+#[repr(C)]
+#[derive(Debug)]
+// TODO: move to cpython/dictobject.rs
+// Not moved because dict.rs uses PyDictObject extensively.
+pub struct PyDictObject {
+    pub ob_base: PyObject,
+    pub ma_used: Py_ssize_t,
+    pub ma_version_tag: u64,
+    pub ma_keys: *mut PyDictKeysObject,
+    pub ma_values: *mut *mut PyObject,
+}
+
+// TODO: move to cpython/dictobject.rs
+opaque_struct!(PyDictKeysObject);
