+++ src/ffi/floatobject.rs
@@ -1,6 +1,8 @@
 use crate::ffi::object::*;
 use std::os::raw::{c_double, c_int};
 
+// TODO: mark non-limited
+// currently used by types/floatob.rs
 #[repr(C)]
 pub struct PyFloatObject {
     pub ob_base: PyObject,
@@ -23,11 +25,8 @@ pub unsafe fn PyFloat_CheckExact(op: *mut PyObject) -> c_int {
     (Py_TYPE(op) == &mut PyFloat_Type) as c_int
 }
 
-#[cfg(not(Py_LIMITED_API))]
-#[inline]
-pub unsafe fn PyFloat_AS_DOUBLE(op: *mut PyObject) -> c_double {
-    (*(op as *mut PyFloatObject)).ob_fval
-}
+// skipped Py_RETURN_NAN
+// skipped Py_RETURN_INF
 
 extern "C" {
     pub fn PyFloat_GetMax() -> c_double;
@@ -40,3 +39,18 @@ extern "C" {
     #[cfg_attr(PyPy, link_name = "PyPyFloat_AsDouble")]
     pub fn PyFloat_AsDouble(arg1: *mut PyObject) -> c_double;
 }
+
+#[cfg(not(Py_LIMITED_API))]
+#[inline]
+pub unsafe fn PyFloat_AS_DOUBLE(op: *mut PyObject) -> c_double {
+    (*(op as *mut PyFloatObject)).ob_fval
+}
+
+// skipped non-limited _PyFloat_Pack2
+// skipped non-limited _PyFloat_Pack4
+// skipped non-limited _PyFloat_Pack8
+// skipped non-limited _PyFloat_Unpack2
+// skipped non-limited _PyFloat_Unpack4
+// skipped non-limited _PyFloat_Unpack8
+// skipped non-limited _PyFloat_DebugMallocStats
+// skipped non-limited _PyFloat_FormatAdvancedWriter
